move modal and textinputmodal to components
